# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API on Kredit Plus Microservice.",
        "title": "API Gateway Kredit Plus",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/api/v1",
    "paths": {
        "/konsumen": {
            "post": {
                "description": "Create new konsumen item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kredit"
                ],
                "summary": "Create konsumen",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateKonsumenRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateKonsumenResponseDto"
                        }
                    }
                }
            }
        },
        "/limit": {
            "post": {
                "description": "Create new limit item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kredit"
                ],
                "summary": "Create limit",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLimitRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLimitResponseDto"
                        }
                    }
                }
            }
        },
        "/limit/{id}": {
            "get": {
                "description": "Get limit by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kredit"
                ],
                "summary": "Get limit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id Limit",
                        "name": "id_limit",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id Konsumen",
                        "name": "id_konsumen",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetLimitResponseDto"
                        }
                    }
                }
            }
        },
        "/transaksi": {
            "post": {
                "description": "Create new transaksi item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kredit"
                ],
                "summary": "Create transaksi",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTransaksiRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTransaksiResponseDto"
                        }
                    }
                }
            }
        },
        "/transaksi/{id}": {
            "get": {
                "description": "Get transaksi by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kredit"
                ],
                "summary": "Get transaksi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id Transaksi",
                        "name": "id_transaksi",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id Konsumen",
                        "name": "id_konsumen",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetTransaksiResponseDto"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateKonsumenRequestDto": {
            "type": "object",
            "required": [
                "email",
                "foto_ktp",
                "foto_selfie",
                "full_name",
                "gaji",
                "id_konsumen",
                "legal_name",
                "nik",
                "password",
                "tanggal_lahir",
                "tempat_lahir"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "foto_ktp": {
                    "type": "string"
                },
                "foto_selfie": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "gaji": {
                    "type": "number",
                    "minimum": 0
                },
                "id_konsumen": {
                    "type": "string"
                },
                "legal_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "nik": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "password": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "tanggal_lahir": {
                    "type": "string"
                },
                "tempat_lahir": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                }
            }
        },
        "dto.CreateKonsumenResponseDto": {
            "type": "object",
            "required": [
                "id_konsumen"
            ],
            "properties": {
                "id_konsumen": {
                    "type": "string"
                }
            }
        },
        "dto.CreateLimitRequestDto": {
            "type": "object",
            "required": [
                "batas_kredit",
                "id_konsumen",
                "id_limit",
                "tenor"
            ],
            "properties": {
                "batas_kredit": {
                    "type": "number",
                    "minimum": 0
                },
                "id_konsumen": {
                    "type": "string"
                },
                "id_limit": {
                    "type": "string"
                },
                "tenor": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "dto.CreateLimitResponseDto": {
            "type": "object",
            "required": [
                "id_limit"
            ],
            "properties": {
                "id_limit": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTransaksiRequestDto": {
            "type": "object",
            "required": [
                "admin_fee",
                "id_konsumen",
                "id_transaksi",
                "jenis_transaksi",
                "jumlah_bunga",
                "jumlah_cicilan",
                "nama_asset",
                "nomor_kontrak",
                "otr",
                "tanggal_transaksi"
            ],
            "properties": {
                "admin_fee": {
                    "type": "number",
                    "minimum": 0
                },
                "id_konsumen": {
                    "type": "string"
                },
                "id_transaksi": {
                    "type": "string"
                },
                "jenis_transaksi": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "jumlah_bunga": {
                    "type": "number",
                    "minimum": 0
                },
                "jumlah_cicilan": {
                    "type": "number",
                    "minimum": 0
                },
                "nama_asset": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "nomor_kontrak": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "otr": {
                    "type": "integer"
                },
                "tanggal_transaksi": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTransaksiResponseDto": {
            "type": "object",
            "required": [
                "id_transaksi"
            ],
            "properties": {
                "id_transaksi": {
                    "type": "string"
                }
            }
        },
        "dto.GetLimitResponseDto": {
            "type": "object",
            "required": [
                "batas_kredit",
                "id_konsumen",
                "id_limit",
                "tenor"
            ],
            "properties": {
                "batas_kredit": {
                    "type": "number",
                    "minimum": 0
                },
                "id_konsumen": {
                    "type": "string"
                },
                "id_limit": {
                    "type": "string"
                },
                "tenor": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "dto.GetTransaksiResponseDto": {
            "type": "object",
            "required": [
                "admin_fee",
                "id_konsumen",
                "id_transaksi",
                "jenis_transaksi",
                "jumlah_bunga",
                "jumlah_cicilan",
                "nama_asset",
                "nomor_kontrak",
                "otr",
                "tanggal_transaksi"
            ],
            "properties": {
                "admin_fee": {
                    "type": "number",
                    "minimum": 0
                },
                "id_konsumen": {
                    "type": "string"
                },
                "id_transaksi": {
                    "type": "string"
                },
                "jenis_transaksi": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "jumlah_bunga": {
                    "type": "number",
                    "minimum": 0
                },
                "jumlah_cicilan": {
                    "type": "number",
                    "minimum": 0
                },
                "nama_asset": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "nomor_kontrak": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 0
                },
                "otr": {
                    "type": "integer"
                },
                "tanggal_transaksi": {
                    "type": "string"
                }
            }
        }
    }
}